Add delete User route
